+++ src/stock-time-series/dto/daily-adjusted-response.dto.ts
@@ -1,0 +1,23 @@
+export type DailyAdjustedMetadata = {
+  information: string;
+  symbol: string;
+  lastRefreshed: string;
+  outputSize: string;
+  timeZone: string;
+};
+
+export type DailyAdjustedTimeSeries = {
+  open: string;
+  high: string;
+  low: string;
+  close: string;
+  volume: string;
+  adjustedClose: string;
+  dividendAmount: string;
+  splitCoefficient: string;
+};
+
+export type DailyAdjustedResponseDTO = {
+  metadata: DailyAdjustedMetadata;
+  timeSeries: { [key: string]: DailyAdjustedTimeSeries };
+};