+++ src/stock-time-series/dto/monthly-adjusted-response.dto.ts
@@ -1,0 +1,21 @@
+export type MonthlyAdjustedMetadata = {
+  information: string;
+  symbol: string;
+  lastRefreshed: string;
+  timeZone: string;
+};
+
+export type MonthlyAdjustedTimeSeries = {
+  open: string;
+  high: string;
+  low: string;
+  close: string;
+  adjustedClose: string;
+  volume: string;
+  dividendAmount: string;
+};
+
+export type MonthlyAdjustedResponseDTO = {
+  metadata: MonthlyAdjustedMetadata;
+  timeSeries: { [key: string]: MonthlyAdjustedTimeSeries };
+};