+++ src/stock-time-series/dto/weekly-adjusted-response.dto.ts
@@ -1,0 +1,21 @@
+export type WeeklyAdjustedMetadata = {
+  information: string;
+  symbol: string;
+  lastRefreshed: string;
+  timeZone: string;
+};
+
+export type WeeklyAdjustedTimeSeries = {
+  open: string;
+  high: string;
+  low: string;
+  close: string;
+  adjustedClose: string;
+  volume: string;
+  dividendAmount: string;
+};
+
+export type WeeklyAdjustedResponseDTO = {
+  metadata: WeeklyAdjustedMetadata;
+  timeSeries: { [key: string]: WeeklyAdjustedTimeSeries };
+};